pyproject.toml
[tool.poetry]
name = "ia-recommendation"
version = "0.1.0"
description = "Sistema de Recomendação com Crew AI"
authors = ["Jessie <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.1"
crewai = "^0.121.0"
langchain-community = "0.2.5" 
langchain = "^0.2.0"
pydantic = "^2.4.2"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
alembic = "^1.16.0"
pydantic-settings = "^2.9.1"
ollama = "^0.1.0"
pytest = "^8.3.3" 
httpx = "^0.27.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.1.0"
pysqlite3-binary = "^0.5.4"
email-validator = "^2.1.1"
langchain-core = "^0.2.7"
chromadb = "0.5.23"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
python-dotenv = "^1.1.0"
pysqlite3-binary = "^0.5.4"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"